UPDATE: Refactored each gui "view" into a separate file, so it is less cluttered and not a monolithic file
