+++ src/gui/DecksListView.tsx
@@ -1,0 +1,170 @@
+import { Platform } from "obsidian";
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+import h from "vhtml";
+
+import type SRPlugin from "src/main";
+import { SRSettings } from "src/settings";
+import { COLLAPSE_ICON } from "src/constants";
+import { t } from "src/lang/helpers";
+import { Deck } from "../Deck";
+import {
+    DeckStats,
+    IFlashcardReviewSequencer as IFlashcardReviewSequencer,
+} from "src/FlashcardReviewSequencer";
+import { TopicPath } from "src/TopicPath";
+import { FlashcardModalMode } from "./flashcard-modal";
+
+export class DecksListView {
+    public plugin: SRPlugin;
+    public mode: FlashcardModalMode;
+    public titleEl: HTMLElement;
+    public contentEl: HTMLElement;
+    private reviewSequencer: IFlashcardReviewSequencer;
+    private settings: SRSettings;
+    private startReviewOfDeck: (deck: Deck) => void;
+
+    constructor(
+        plugin: SRPlugin,
+        settings: SRSettings,
+        reviewSequencer: IFlashcardReviewSequencer,
+        titleEl: HTMLElement,
+        contentEl: HTMLElement,
+        startReviewOfDeck: (deck: Deck) => void
+    ) {
+        this.plugin = plugin;
+        this.settings = settings;
+        this.reviewSequencer = reviewSequencer;
+
+        this.startReviewOfDeck = startReviewOfDeck;
+
+        this.titleEl = titleEl;
+        this.contentEl = contentEl;
+
+        this.titleEl.addClass("sr-centered");
+
+        this.contentEl.style.position = "relative";
+        this.contentEl.style.height = "92%";
+        this.contentEl.addClass("sr-modal-content");
+        if (Platform.isMobile) {
+            this.contentEl.style.display = "block";
+        }
+    }
+    /**
+     * Shows the DeckListView
+     */
+    show(): void {
+        this.mode = FlashcardModalMode.DecksList;
+
+        const stats: DeckStats = this.reviewSequencer.getDeckStats(TopicPath.emptyPath);
+
+        this.titleEl.setText(t("DECKS"));
+        this.titleEl.innerHTML += (
+            <p style="margin:0px;line-height:12px;">
+                <span
+                    style="background-color:#4caf50;color:#ffffff;"
+                    aria-label={t("DUE_CARDS")}
+                    class="tag-pane-tag-count tree-item-flair sr-deck-counts"
+                >
+                    {stats.dueCount.toString()}
+                </span>
+                <span
+                    style="background-color:#2196f3;"
+                    aria-label={t("NEW_CARDS")}
+                    class="tag-pane-tag-count tree-item-flair sr-deck-counts"
+                >
+                    {stats.newCount.toString()}
+                </span>
+                <span
+                    style="background-color:#ff7043;"
+                    aria-label={t("TOTAL_CARDS")}
+                    class="tag-pane-tag-count tree-item-flair sr-deck-counts"
+                >
+                    {stats.totalCount.toString()}
+                </span>
+            </p>
+        );
+        this.contentEl.empty();
+        this.contentEl.setAttribute("id", "sr-flashcard-view");
+
+        for (const deck of this.reviewSequencer.originalDeckTree.subdecks) {
+            this.renderDeck(deck, this.contentEl);
+        }
+    }
+
+    renderDeck(deck: Deck, containerEl: HTMLElement): void {
+        const deckView: HTMLElement = containerEl.createDiv("tree-item");
+
+        const deckViewSelf: HTMLElement = deckView.createDiv(
+            "tree-item-self tag-pane-tag is-clickable",
+        );
+        const shouldBeInitiallyExpanded: boolean = this.settings.initiallyExpandAllSubdecksInTree;
+        let collapsed = !shouldBeInitiallyExpanded;
+        let collapseIconEl: HTMLElement | null = null;
+        if (deck.subdecks.length > 0) {
+            collapseIconEl = deckViewSelf.createDiv("tree-item-icon collapse-icon");
+            collapseIconEl.innerHTML = COLLAPSE_ICON;
+            (collapseIconEl.childNodes[0] as HTMLElement).style.transform = collapsed
+                ? "rotate(-90deg)"
+                : "";
+        }
+
+        const deckViewInner: HTMLElement = deckViewSelf.createDiv("tree-item-inner");
+        const deckViewInnerText: HTMLElement = deckViewInner.createDiv("tag-pane-tag-text");
+        deckViewInnerText.innerHTML += <span class="tag-pane-tag-self">{deck.deckName}</span>;
+        const deckViewOuter: HTMLElement = deckViewSelf.createDiv("tree-item-flair-outer");
+        const deckStats = this.reviewSequencer.getDeckStats(deck.getTopicPath());
+        deckViewOuter.innerHTML += (
+            <span>
+                <span
+                    style="background-color:#4caf50;"
+                    class="tag-pane-tag-count tree-item-flair sr-deck-counts"
+                >
+                    {deckStats.dueCount.toString()}
+                </span>
+                <span
+                    style="background-color:#2196f3;"
+                    class="tag-pane-tag-count tree-item-flair sr-deck-counts"
+                >
+                    {deckStats.newCount.toString()}
+                </span>
+                <span
+                    style="background-color:#ff7043;"
+                    class="tag-pane-tag-count tree-item-flair sr-deck-counts"
+                >
+                    {deckStats.totalCount.toString()}
+                </span>
+            </span>
+        );
+
+        const deckViewChildren: HTMLElement = deckView.createDiv("tree-item-children");
+        deckViewChildren.style.display = collapsed ? "none" : "block";
+        if (deck.subdecks.length > 0) {
+            collapseIconEl.addEventListener("click", (e) => {
+                if (collapsed) {
+                    (collapseIconEl.childNodes[0] as HTMLElement).style.transform = "";
+                    deckViewChildren.style.display = "block";
+                } else {
+                    (collapseIconEl.childNodes[0] as HTMLElement).style.transform =
+                        "rotate(-90deg)";
+                    deckViewChildren.style.display = "none";
+                }
+
+                // We stop the propagation of the event so that the click event for deckViewSelf doesn't get called
+                // if the user clicks on the collapse icon
+                e.stopPropagation();
+                collapsed = !collapsed;
+            });
+        }
+
+        // Add the click handler to deckViewSelf instead of deckViewInner so that it activates
+        // over the entire rectangle of the tree item, not just the text of the topic name
+        // https://github.com/st3v3nmw/obsidian-spaced-repetition/issues/709
+        deckViewSelf.addEventListener("click", () => {
+            this.startReviewOfDeck(deck);
+        });
+
+        for (const subdeck of deck.subdecks) {
+            this.renderDeck(subdeck, deckViewChildren);
+        }
+    }
+}
